build: Fix clang warnings
Latest clang generated multiple build warnings.  This should resolve
the issues encountered.

+++ layers/gpu_utils.cpp
@@ -449,7 +449,7 @@ bool GetLineAndFilename(const std::string string, uint32_t *linenumber, std::str
         }
     }
     if (0 == line_index) return false;
-    *linenumber = std::stoul(tokens[line_index]);
+    *linenumber = static_cast<uint32_t>(std::stoul(tokens[line_index]));
     uint32_t filename_index = line_index + 1;
     // Remove enclosing double quotes around filename
     if (size > filename_index) filename = tokens[filename_index].substr(1, tokens[filename_index].size() - 2);
